Fix migration destination replica index bug
* Migration destination replica index should be set properly on MigrationOperation so that migration tasks can access it correctly.
* BaseMigrationSourceOperation and BaseMigrationDestinationOperation were being as base classes for migration classes of 3.8 and 3.9. They are now deleted because they are not needed anymore.
