+++ hazelcast/src/main/java/com/hazelcast/internal/partition/operation/MigrationRequestOperation.java
@@ -16,12 +16,15 @@
 
 package com.hazelcast.internal.partition.operation;
 
+import com.hazelcast.core.Member;
+import com.hazelcast.core.MemberLeftException;
 import com.hazelcast.internal.partition.InternalPartition;
 import com.hazelcast.internal.partition.InternalPartitionService;
 import com.hazelcast.internal.partition.MigrationInfo;
 import com.hazelcast.internal.partition.NonFragmentedServiceNamespace;
 import com.hazelcast.internal.partition.PartitionReplicaVersionManager;
 import com.hazelcast.internal.partition.ReplicaFragmentMigrationState;
+import com.hazelcast.internal.partition.impl.InternalMigrationListener.MigrationParticipant;
 import com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl;
 import com.hazelcast.internal.partition.impl.MigrationManager;
 import com.hazelcast.internal.partition.impl.PartitionDataSerializerHook;
@@ -29,17 +32,22 @@ import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
+import com.hazelcast.spi.ExceptionAction;
 import com.hazelcast.spi.FragmentedMigrationAwareService;
 import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.spi.Operation;
+import com.hazelcast.spi.PartitionMigrationEvent;
 import com.hazelcast.spi.PartitionReplicationEvent;
 import com.hazelcast.spi.ServiceNamespace;
 import com.hazelcast.spi.UrgentSystemOperation;
+import com.hazelcast.spi.exception.RetryableHazelcastException;
+import com.hazelcast.spi.exception.TargetNotMemberException;
 import com.hazelcast.spi.impl.NodeEngineImpl;
 import com.hazelcast.spi.impl.PartitionSpecificRunnable;
 import com.hazelcast.spi.impl.SimpleExecutionCallback;
 import com.hazelcast.spi.impl.operationservice.InternalOperationService;
 import com.hazelcast.spi.impl.servicemanager.ServiceInfo;
+import com.hazelcast.spi.partition.MigrationEndpoint;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -49,6 +57,8 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ExecutionException;
+import java.util.logging.Level;
 
 import static java.util.Collections.singleton;
 
@@ -59,8 +69,9 @@ import static java.util.Collections.singleton;
  * It runs on the migration source and transfers the partition with multiple shots.
  * It divides the partition data into fragments and send a group of fragments within each shot.
  */
-public class MigrationRequestOperation extends BaseMigrationSourceOperation {
+public class MigrationRequestOperation extends BaseMigrationOperation {
 
+    private boolean returnResponse = true;
     private boolean fragmentedMigrationEnabled;
     private transient ServiceNamespacesContext namespacesContext;
 
@@ -71,6 +82,107 @@ public class MigrationRequestOperation extends BaseMigrationSourceOperation {
                                      boolean fragmentedMigrationEnabled) {
         super(migrationInfo, partitionStateVersion);
         this.fragmentedMigrationEnabled = fragmentedMigrationEnabled;
+    }
+
+    @Override
+    protected PartitionMigrationEvent getMigrationEvent() {
+        return new PartitionMigrationEvent(MigrationEndpoint.SOURCE,
+                migrationInfo.getPartitionId(), migrationInfo.getSourceCurrentReplicaIndex(),
+                migrationInfo.getSourceNewReplicaIndex());
+    }
+
+    @Override
+    protected MigrationParticipant getMigrationParticipantType() {
+        return MigrationParticipant.SOURCE;
+    }
+
+    /** Verifies that the local master is equal to the migration master. */
+    private void verifyMasterOnMigrationSource() {
+        NodeEngine nodeEngine = getNodeEngine();
+        Address masterAddress = nodeEngine.getMasterAddress();
+
+        if (!migrationInfo.getMaster().equals(masterAddress)) {
+            throw new IllegalStateException("Migration initiator is not master node! => " + toString());
+        }
+
+        if (!masterAddress.equals(getCallerAddress())) {
+            throw new IllegalStateException("Caller is not master node! => " + toString());
+        }
+    }
+
+    /** Verifies that this node is the owner of the partition. */
+    private void verifySource(Address thisAddress, InternalPartition partition) {
+        Address owner = partition.getOwnerOrNull();
+        if (owner == null) {
+            throw new RetryableHazelcastException("Cannot migrate at the moment! Owner of the partition is null => "
+                    + migrationInfo);
+        }
+
+        if (!thisAddress.equals(owner)) {
+            throw new RetryableHazelcastException("Owner of partition is not this node! => " + toString());
+        }
+    }
+
+    /** Verifies that the destination is a cluster member. */
+    private void verifyExistingTarget(NodeEngine nodeEngine, Address destination) {
+        Member target = nodeEngine.getClusterService().getMember(destination);
+        if (target == null) {
+            throw new TargetNotMemberException("Destination of migration could not be found! => " + toString());
+        }
+    }
+
+    private PartitionReplicationEvent getPartitionReplicationEvent() {
+        return new PartitionReplicationEvent(migrationInfo.getPartitionId(), migrationInfo.getDestinationNewReplicaIndex());
+    }
+
+    private void setFailed() {
+        success = false;
+        onMigrationComplete(false);
+    }
+
+    private void completeMigration(boolean result) {
+        success = result;
+        migrationInfo.doneProcessing();
+        onMigrationComplete(result);
+        sendResponse(result);
+    }
+
+    private void logThrowable(Throwable t) {
+        Throwable throwableToLog = t;
+        if (throwableToLog instanceof ExecutionException) {
+            throwableToLog = throwableToLog.getCause() != null ? throwableToLog.getCause() : throwableToLog;
+        }
+        Level level = getLogLevel(throwableToLog);
+        getLogger().log(level, throwableToLog.getMessage(), throwableToLog);
+    }
+
+    private Level getLogLevel(Throwable e) {
+        return (e instanceof MemberLeftException || e instanceof InterruptedException)
+                || !getNodeEngine().isRunning() ? Level.INFO : Level.WARNING;
+    }
+
+    @Override
+    public ExceptionAction onInvocationException(Throwable throwable) {
+        if (throwable instanceof TargetNotMemberException) {
+            return ExceptionAction.THROW_EXCEPTION;
+        }
+        return super.onInvocationException(throwable);
+    }
+
+    @Override
+    public boolean returnsResponse() {
+        return returnResponse;
+    }
+
+    @Override
+    void executeBeforeMigrations() throws Exception {
+        NodeEngine nodeEngine = getNodeEngine();
+        boolean ownerMigration = nodeEngine.getThisAddress().equals(migrationInfo.getSource());
+        if (!ownerMigration) {
+            return;
+        }
+
+        super.executeBeforeMigrations();
     }
 
     @Override
@@ -134,7 +246,7 @@ public class MigrationRequestOperation extends BaseMigrationSourceOperation {
      * Invokes the {@link MigrationOperation} on the migration destination.
      */
     private void invokeMigrationOperation(Address destination, ReplicaFragmentMigrationState migrationState,
-                                          boolean firstFragment) throws IOException {
+                                          boolean firstFragment) {
 
         boolean lastFragment = !fragmentedMigrationEnabled || !namespacesContext.hasNext();
         Operation operation = new MigrationOperation(migrationInfo, partitionStateVersion, migrationState,
@@ -158,7 +270,6 @@ public class MigrationRequestOperation extends BaseMigrationSourceOperation {
                 .setCallTimeout(partitionService.getPartitionMigrationTimeout())
                 .setTryCount(InternalPartitionService.MIGRATION_RETRY_COUNT)
                 .setTryPauseMillis(InternalPartitionService.MIGRATION_RETRY_PAUSE)
-                .setReplicaIndex(getReplicaIndex())
                 .invoke();
     }
